pyproject.toml
[project]
name = "email-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.0",
    "google-api-python-client>=2.192.0",
    "google-auth>=2.49.0",
    "google-auth-oauthlib>=1.3.0",
    "google-genai>=1.66.0",
    "mcp[cli]>=1.26.0",
    "pypdf2>=3.0.1",
    "streamlit>=1.55.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]
